% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{bind_cols}
\alias{bind_cols}
\title{Bind_cols to manage map_dfc in package
This function as no use}
\usage{
bind_cols(
  ...,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)
}
\arguments{
\item{...}{ok}

\item{.name_repair}{ok}
}
\value{
bind rows
}
\description{
Bind_cols to manage map_dfc in package
This function as no use
}
