% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{chr_print_buffed}
\alias{chr_print_buffed}
\title{Get character encoding the list or data.fraùe hard writing}
\usage{
chr_print_buffed(x, type = "list")
}
\arguments{
\item{x}{character}

\item{type}{'list' or 'description'data.frame', what is the wanted results}
}
\value{
a chr
}
\description{
Get character encoding the list or data.fraùe hard writing
}
