% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{copy_to_clipboard
Safe wrapper for clipr::write_clip. I want to use it but it doesn't pass checks correctly}
\usage{
copy_to_clipboard(x)
}
\arguments{
\item{x}{string to copy to clip board}
}
\value{
clipr::write_clip(x) if it can
}
\description{
copy_to_clipboard
Safe wrapper for clipr::write_clip. I want to use it but it doesn't pass checks correctly
}
