% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{get_chr_to_quote}
\alias{get_chr_to_quote}
\title{get_chr_to_quote}
\usage{
get_chr_to_quote(x)
}
\arguments{
\item{x}{character}
}
\value{
a boolean vector wether or not each element need to be strong quoted
}
\description{
This function identify weird names format and correct them with : ``
}
