% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{print_for_r}
\alias{print_for_r}
\title{create character describing x as you would write it manually in R.}
\usage{
print_for_r(x, shape = NULL, add_space = 0)
}
\arguments{
\item{x}{a vector/list/data.frame/matrix to write manually}

\item{shape}{(default : `NULL`) decide what shape to print: only when x is a vector and not a list}

\item{add_space}{for recurcive use of function keep nb of " " tabulation 
- vector -> "vector"
- dyplr::select -> "select"}
}
\description{
create character describing x as you would write it manually in R.
}
