% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{rclip}
\alias{rclip}
\title{Add to clipboard text describing x as you would write it manually in R.}
\usage{
rclip(x, shape = NULL)
}
\arguments{
\item{x}{a vector/list/data.frame/matrix to write manually}

\item{shape}{(default : `NULL`) decide what shape to print: only when x is a vector and not a list
- vector -> "vector"
- dyplr::select -> "select"}
}
\description{
Mainly usefull when i need to hard code a set of values that are to be extracted. 
Example: I want to use all columns names of a data.frame as a vector
}
\examples{

# To use it extract the vector you want to hard code somewhere

data <- sample(x=c("ceci","est","un","test"), size = 1000, replace = TRUE) \%>\% 
  as.factor() \%>\% 
   summary()
   
vec_to_hard_code <- which(data > 250) \%>\% 
   names()
   
rclip(vec_to_hard_code)
rclip(vec_to_hard_code,shape = 'select')

list_to_hard_code <- list(a = sample(1:5,size = 5),
                                b = sample(1:5,size = 10,
                                replace = TRUE))
rclip(list_to_hard_code)

data_to_hard_code <- data.frame(a = sample(1:10,size = 10),
                                b = sample(1:5,size = 10,
                                replace = TRUE))
rclip(data_to_hard_code)

rclip(as.matrix(data_to_hard_code))
# ready to paste ! Press Ctrl + V 

}
