% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming_files.R
\name{renumerate_files}
\alias{renumerate_files}
\title{Renumerated files and folders names in path}
\usage{
renumerate_files(path = ".", sep = "_", check = TRUE, rename = TRUE)
}
\arguments{
\item{path}{Location for folders and files to rename}

\item{sep}{separation character between numeration and actual name}

\item{check}{Boolean asks user validation ?}

\item{rename}{Boolean Actually rename ?}
}
\value{
Can return a character vector of renamed files if rename = FALSE
}
\description{
Renumerated files and folders names in path
}
\examples{
renumerate_files(path = ".",sep ='_', check = TRUE, rename = FALSE)
\dontrun{
renumerate_files(path = "my_files_adress",sep =' - ')
}
}
