% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming_files.R
\name{renumerate_folders}
\alias{renumerate_folders}
\title{Renumerated files and/or folders names in path}
\usage{
renumerate_folders(path = ".", ...)
}
\arguments{
\item{path}{Location for folders and files to rename}

\item{...}{:  Additionnal parameters : 
- `folders = TRUE`: should it rename folders
- `files = FALSE`: should it rename files
- `sep = "_"` : separation character between numeration and actual name
- `hidden = TRUE`: Consider hidden files/folders
- `check = TRUE`: Boolean asks for user validation ?
- `rename = TRUE`: Boolean Actually rename ?}
}
\value{
Can return a character vector of renamed files if rename = FALSE
}
\description{
Renumerated files and/or folders names in path
}
\examples{
renumerate_folders(path = ".", rename = FALSE)
\dontrun{
renumerate_folders(path = "my_files_adress", 
                   sep = " - ", files = TRUE, 
                   folders = FALSE)
}
}
