% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{chr_print_buffed}
\alias{chr_print_buffed}
\title{Get character encoding the list or data.fraùe hard writing}
\usage{
chr_print_buffed(x, type = "list", add_space = 0)
}
\arguments{
\item{x}{list or data.frame}

\item{type}{'list' or 'description'data.frame', what is the wanted results}

\item{add_space}{for recurcive use of function keep nb of " " tabulation}
}
\value{
a chr
}
\description{
Get character encoding the list or data.fraùe hard writing
}
