% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{chr_print_vector}
\alias{chr_print_vector}
\title{Get character encoding the vector hard writing}
\usage{
chr_print_vector(x)
}
\arguments{
\item{x}{character}
}
\value{
a chr
}
\description{
Get character encoding the vector hard writing
}
