% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming_files.R
\name{enumerated_files_names}
\alias{enumerated_files_names}
\title{Gives a character vector of new numerated names for files in path}
\usage{
enumerated_files_names(path = ".", sep = "_")
}
\arguments{
\item{path}{Location for folders and files to rename}

\item{sep}{separation character between numeration and actual name}
}
\value{
a character with all new names
}
\description{
Gives a character vector of new numerated names for files in path
}
