% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_copy_paste_vector.R
\name{print_to_copy}
\alias{print_to_copy}
\title{Print a vector as you would write it manually.}
\usage{
print_to_copy(vector)
}
\arguments{
\item{vector}{to write manually}
}
\description{
Mainly usefull when i need to hard code a set of values that are to be extracted. 
Example: I want to use all columns names of a data.frame as a vector
}
\examples{

# To use it extract the vector you want to hard code somewhere

data <- sample(x=c("ceci","est","un","test"), size = 1000, replace = TRUE) \%>\% 
  as.factor() \%>\% 
   summary()
   
vec_to_hard_code <- which(data > 250) \%>\% 
   names()
   
print_to_copy(vec_to_hard_code)


}
