% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming_files.R
\name{renumerate_folders}
\alias{renumerate_folders}
\title{Renumerated files and/or folders names in path}
\usage{
renumerate_folders(path = ".", ...)
}
\arguments{
\item{path}{Location for folders and files to rename}

\item{...}{Additionnal parameters : 
\describe{
    \item{folders}{TRUE, should it rename folders}
    \item{files}{FALSE, should it rename files}
    \item{sep}{"_", separation character between numeration and actual name}
    \item{hidden}{TRUE, Consider hidden files/folders ?}
    \item{check}{TRUE, Boolean asks for user validation ?}
    \item{rename}{TRUE,  Boolean Actually rename ?}
    \item{without}{character vector, grep like pattern to be excluded from enumeration}
  }}
}
\value{
Can return a character vector of renamed files if rename = FALSE
}
\description{
Renumerated files and/or folders names in path
}
\examples{
renumerate_folders(path = ".", rename = FALSE)
\dontrun{
renumerate_folders(path = "my_files_adress", 
                   sep = " - ", files = TRUE, 
                   folders = FALSE)
}
}
